refactor: confirmation_dialog.dart -> dialogs.dart
