+++ lib/src/widgets/release_card.dart
@@ -6,7 +6,7 @@ import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
 import '../../fwupd_x.dart';
-import 'confirmation_dialog.dart';
+import 'dialogs.dart';
 
 class ReleaseCard extends StatelessWidget {
   const ReleaseCard({
